Set config reader for docker secrets

+++ database/database.go
@@ -17,7 +17,7 @@ type Database struct {
 	PostgresHost       string
 	PostgresUser       string
 	PostgresPassword   string
-	PostgresPort       int
+	PostgresPort       int64
 	PostgresDb         string
 	PostgresRequireSSL string
 	Connection         *sql.DB
